Rearranged package structure
