+++ katz/src/main/kotlin/katz/data/Try.kt
