+++ katz/src/main/kotlin/katz/instances/OptionMonad.kt
@@ -16,6 +16,16 @@
 
 package katz
 
-interface Functor<F> {
-    fun <A, B> map(fa: HK<F, A>, f: (A) -> B): HK<F, B>
+object OptionMonad : Monad<Option.F> {
+
+    fun <A> HK<Option.F, A>.ev(): Option<A> = this as Option<A>
+
+    override fun <A, B> map(fa: HK<Option.F, A>, f: (A) -> B): HK<Option.F, B> =
+            fa.ev().map(f)
+
+    override fun <A> pure(a: A): HK<Option.F, A> = Option.Some(a)
+
+    override fun <A, B> flatMap(fa: HK<Option.F, A>, f: (A) -> HK<Option.F, B>): HK<Option.F, B> =
+            fa.ev().flatMap { f(it).ev() }
+
 }
