+++ katz/src/test/kotlin/katz/EitherTest.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2017 The Kats Authors
+ * Copyright (C) 2017 The Katz Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -18,9 +18,9 @@ package katz
 
 import io.kotlintest.KTestJUnitRunner
 import io.kotlintest.properties.forAll
-import katz.Either.Left
-import katz.Either.Right
 import org.junit.runner.RunWith
+
+import katz.Either.*
 
 @RunWith(KTestJUnitRunner::class)
 class EitherTest : UnitSpec() {
